Changed deltaCommission from uint16 to uint8

+++ contracts/Core/parameters/child/StakeManagerParams.sol
@@ -17,7 +17,7 @@ abstract contract StakeManagerParams is ACL, IStakeManagerParams {
     // Slash Penalty = bounty + burned + kept
     uint16 public override baseDenominator = 10000;
     // change the commission by 3% points
-    uint16 public deltaCommission = 3;
+    uint8 public deltaCommission = 3;
     uint16 public gracePeriod = 8;
     uint16 public epochLength = 300;
     bool public escapeHatchEnabled = true;
